Chromedriver service uses port & path from config
This change makes it so the chromedriver service uses the port and path defined in `wdio.conf.js` rather than the other way round. This makes it easier to install and use the service, without having to worry about configuring special configuration specifically for chromedriver.path defined.

For example, when installing the chromedriver service via the `wdio config` wizard, it is not clear that you have to update the path and port in order to make it work. With this change, it would always work regardless of how port and path are defined.